RapidPy/com_port_mapper/requirements.txt
PySide6>=6.8,<6.9
pyserial>=3.5,<4